+++ src/Symfony/Component/Config/Tests/Definition/Builder/TreeBuilderTest.php
@@ -150,4 +150,65 @@ class TreeBuilderTest extends TestCase
 
         $this->assertEquals(array('enabled' => false), $result);
     }
+
+    public function testDefaultPathSeparatorIsDot()
+    {
+        $builder = new TreeBuilder();
+
+        $builder->root('propagation')
+            ->children()
+                ->node('foo', 'variable')->end()
+                ->arrayNode('child')
+                    ->children()
+                        ->node('foo', 'variable')
+                    ->end()
+                ->end()
+            ->end()
+        ->end();
+
+        $node = $builder->buildTree();
+        $children = $node->getChildren();
+
+        $this->assertArrayHasKey('foo', $children);
+        $this->assertInstanceOf('Symfony\Component\Config\Definition\BaseNode', $children['foo']);
+        $this->assertSame('propagation.foo', $children['foo']->getPath());
+
+        $this->assertArrayHasKey('child', $children);
+        $childChildren = $children['child']->getChildren();
+
+        $this->assertArrayHasKey('foo', $childChildren);
+        $this->assertInstanceOf('Symfony\Component\Config\Definition\BaseNode', $childChildren['foo']);
+        $this->assertSame('propagation.child.foo', $childChildren['foo']->getPath());
+    }
+
+    public function testPathSeparatorIsPropagatedToChildren()
+    {
+        $builder = new TreeBuilder();
+
+        $builder->root('propagation')
+            ->children()
+                ->node('foo', 'variable')->end()
+                ->arrayNode('child')
+                    ->children()
+                        ->node('foo', 'variable')
+                    ->end()
+                ->end()
+            ->end()
+        ->end();
+
+        $builder->setPathSeparator('/');
+        $node = $builder->buildTree();
+        $children = $node->getChildren();
+
+        $this->assertArrayHasKey('foo', $children);
+        $this->assertInstanceOf('Symfony\Component\Config\Definition\BaseNode', $children['foo']);
+        $this->assertSame('propagation/foo', $children['foo']->getPath());
+
+        $this->assertArrayHasKey('child', $children);
+        $childChildren = $children['child']->getChildren();
+
+        $this->assertArrayHasKey('foo', $childChildren);
+        $this->assertInstanceOf('Symfony\Component\Config\Definition\BaseNode', $childChildren['foo']);
+        $this->assertSame('propagation/child/foo', $childChildren['foo']->getPath());
+    }
 }